requirements.txt
discord.py==1.3.3
python-dotenv==0.13.0
aiohttp==3.6.2
aiocache==0.11.1
ujson==2.0.3
msgpack==1.0.0
